use 301 status code for directory redirects
The 303 error code isn't really appropriate for this use case. It should be
either a 301 (permanent) or a 302 (temporary). Using either 302 or 303 means
that search engines won't pass along ranking to the target so it's lost. It
also prevents caching of the redirect, which is done for 301. The cache can be
invalidated by redirecting back from the target so it's harmless.